# knowledge-base/knowledge_base/phase9_integration/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WHO ITA Knowledge Base API",
    "description": "REST API for accessing WHO International Standard Terminologies on Ayurveda",
    "version": "1.0.0",
    "contact": {
      "name": "Ayurvritta AIMRS",
      "email": "support@ayurvritta.com"
    }
  },
  "servers": [
    {
      "url": "/api/v1/ita",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/autocomplete": {
      "get": {
        "summary": "Autocomplete suggestions",
        "description": "Get term suggestions for autocomplete",
        "operationId": "autocomplete",
        "parameters": [
          {
            "name": "prefix",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Suggestions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "term_id": {
                        "type": "string"
                      },
                      "english": {
                        "type": "string"
                      },
                      "iast": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}